+++ clients/include/testing_spr_strided_batched.hpp
@@ -28,7 +28,7 @@ void testing_spr_strided_batched_bad_arg()
 
     rocblas_local_handle handle;
 
-    size_t size_A = N * (N + 1) / 2;
+    size_t size_A = size_t(N) * (N + 1) / 2;
 
     // allocate memory on device
     device_strided_batch_vector<T> dA_1(size_A, 1, stride_A, batch_count);
@@ -79,7 +79,7 @@ void testing_spr_strided_batched(const Arguments& arg)
     }
 
     size_t abs_incx = incx >= 0 ? incx : -incx;
-    size_t size_A   = size_t((N * (N + 1)) / 2);
+    size_t size_A   = size_t(N) * (N + 1) / 2;
 
     // Naming: dK is in GPU (device) memory. hK is in CPU (host) memory
     host_strided_batch_vector<T> hA_1(size_A, 1, stride_A, batch_count);
